Add periodic neighbor order function + tests for it.
In order to work with the cell linked neighbor list, one needs to
know which cells (labeled by increasing numbers starting from zero)
are neighbors of each other. The function provides a list of
np arrays that contain all the neighbors. Periodic boundaries are
accounted for and double interactions are excluded, i.e. no
neighbor pair occurs twice in the list.